+++ scripts/gb_model/redispatch/retrieve_bid_offer_data_elexon.py
@@ -71,9 +71,7 @@ async def fetch_api_request_data(
     return df
 
 
-async def get_historical_bod(
-    base_url: str, bmu_carrier_map: dict[str, str], year: int
-) -> pd.DataFrame:
+async def get_historical_bod(base_url: str, year: int) -> pd.DataFrame:
     """
     Fetch bid / offer data for an entire year by querying day by day
 
@@ -124,74 +122,8 @@ async def get_historical_bod(
     logger.info(f"Successfully retrieved Bid/offer price data for {year}.")
 
     df_bod = pd.concat(results, ignore_index=True)
-    df_bod["carrier"] = df_bod["nationalGridBmUnit"].map(bmu_carrier_map)
 
-    df_bod_mean = pd.DataFrame()
-    # BidOfferPairId is an indication of the bandwidth within which a BMunit can increase / decrease it's power output.
-    # -ve pairId's indicate bids and +ve pairId's indicate offers
-    # The bid / offer price can vary with the pairId - for simplicity an average of the prices over the pair id's is used here
-    df_bod.settlementDate = pd.to_datetime(df_bod.settlementDate)
-    df_bod_mean = (
-        df_bod.query("bidOfferPairId < 0")
-        .groupby([pd.Grouper(key="settlementDate", freq="QE"), "carrier"])["bidPrice"]
-        .mean()
-    )
-    offer_price = (
-        df_bod.query("bidOfferPairId > 0")
-        .groupby([pd.Grouper(key="settlementDate", freq="QE"), "carrier"])["offerPrice"]
-        .mean()
-    )
-    df_bod_mean = pd.DataFrame(df_bod_mean).join(pd.DataFrame(offer_price), how="inner")
-
-    return df_bod_mean
-
-
-async def fetch_BM_units(
-    base_url: str,
-    technology_mapping: dict[str, str],
-    bmu_fuel_map_path: str,
-    api_bmu_fuel_map: bool,
-) -> tuple[str, dict]:
-    """
-    Fetch BM Unit data from Elexon API to get mapping of BM units to fuel types
-
-    Parameters
-    ----------
-    base_url: str
-        Base URL for the Elexon BMRS API
-    technology_mapping: dict[str, str]
-        Map Elexon carrier types to PyPSA carrier types
-    bmu_fuel_map_path: str
-        CSV path for mapping of BMU units to their fueltype
-    api_bmu_fuel_map: bool
-        Boolean to choose between fetching BM units via API (True) / reading existing excel (False)
-    """
-
-    if api_bmu_fuel_map:
-        # URL to fetch BM unit data
-        url = f"{base_url}/reference/bmunits/all"
-
-        # Fetch BM unit data
-        async with aiohttp.ClientSession(trust_env=True) as session:
-            df_bmu = await fetch_api_request_data(
-                url, retrieval_message="BMU Unit Data", session=session
-            )
-    else:
-        df_bmu = pd.read_excel(bmu_fuel_map_path).rename(
-            columns={"NESO BMU ID": "nationalGridBmUnit", "REG FUEL TYPE": "fuelType"}
-        )
-
-    # Map BM unit fuel types to PyPSA carriers
-    df_bmu["carrier"] = df_bmu["fuelType"].map(technology_mapping)
-
-    # Filter to only include BM units relevant to conventional technologies in the model
-    df_bmu_filtered = df_bmu.loc[df_bmu["carrier"].isin(technology_mapping.values())]
-
-    # Dictionary to map BM units to carriers
-    bmu_carrier_map = dict(df_bmu_filtered[["nationalGridBmUnit", "carrier"]].values)
-
-    logger.info(f"Created BM unit mapping to carriers {technology_mapping.keys()}.")
-    return bmu_carrier_map
+    return df_bod
 
 
 if __name__ == "__main__":
@@ -215,19 +147,9 @@ if __name__ == "__main__":
     global SEM
     SEM = asyncio.Semaphore(max_concurrent_requests)
 
-    bmu_carrier_map = asyncio.run(
-        fetch_BM_units(
-            base_url,
-            technology_mapping=snakemake.params.technology_mapping,
-            bmu_fuel_map_path=snakemake.input.bmu_fuel_map_path,
-            api_bmu_fuel_map=snakemake.params.api_bmu_fuel_map,
-        )
-    )
-
     df_bod = asyncio.run(
         get_historical_bod(
             base_url,
-            bmu_carrier_map,
             year=int(snakemake.wildcards.bod_year),
         )
     )